Project/MySQL격자포인트/requirements.txt
# Database connectivity
mysql-connector-python==8.0.33
psycopg2-binary==2.9.6

# Environment variable management
python-dotenv==1.0.0
